// src/api/typescript-rxjs/models/Playlist.ts
// tslint:disable
/**
 * Spotify Web API
 * Welcome to the _Spotify Web API_ specification. For more information about how to use the API, check out [Spotify\'s developer site](https://developer.spotify.com/web-api/). 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    Image,
    PlaylistFollowers,
    PlaylistTrackPage,
    UserProfile,
} from './';

/**
 * @export
 * @interface Playlist
 */
export interface Playlist {
    /**
     * True if the owner allows other users to modify the playlist.
     * @type {boolean}
     * @memberof Playlist
     */
    collaborative?: boolean;
    /**
     * The playlist description. Only returned for modified, verified playlists, otherwise null.
     * @type {string}
     * @memberof Playlist
     */
    description?: string;
    /**
     * Known external URLs for this album.
     * @type {{ [key: string]: string; }}
     * @memberof Playlist
     */
    external_urls?: { [key: string]: string; };
    /**
     * @type {PlaylistFollowers}
     * @memberof Playlist
     */
    followers?: PlaylistFollowers;
    /**
     * A link to the Web API endpoint providing full details of the playlist.
     * @type {string}
     * @memberof Playlist
     */
    href?: string;
    /**
     * The Spotify ID of the playlist.
     * @type {string}
     * @memberof Playlist
     */
    id?: string;
    /**
     * The cover art for the album in various sizes, widest first.
     * @type {Array<Image>}
     * @memberof Playlist
     */
    images?: Array<Image>;
    /**
     * The name of the playlist.
     * @type {string}
     * @memberof Playlist
     */
    name?: string;
    /**
     * @type {UserProfile}
     * @memberof Playlist
     */
    owner?: UserProfile;
    /**
     * The playlist\'s public/private status: true the playlist is public, false the playlist is private, null the playlist status is not relevant. For more about public/private status, see Working with Playlists.
     * @type {boolean}
     * @memberof Playlist
     */
    _public?: boolean;
    /**
     * The version identifier for the current playlist. Can be supplied in other requests to target a specific playlist version
     * @type {string}
     * @memberof Playlist
     */
    snapshot_id?: string;
    /**
     * @type {PlaylistTrackPage}
     * @memberof Playlist
     */
    tracks?: PlaylistTrackPage;
    /**
     * The object type: \'playlist\'.
     * @type {string}
     * @memberof Playlist
     */
    type?: string;
    /**
     * Spotify URI of the playlist.
     * @type {string}
     * @memberof Playlist
     */
    uri?: string;
}